requirements.txt
modelscope==1.9.5
streamlit==1.32.0
transformers==4.36.2
sentencepiece==0.1.99
accelerate==0.24.1
transformers_stream_generator==0.0.4
safetensors
bitsandbytes
streamlit_image_select
streamlit_modal
tencentcloud-sdk-python
pillow